main.py
import os
import uvicorn
from google.adk.cli.fast_api import get_fast_api_app

app = get_fast_api_app()

if __name__ == "__main__":
    port = int(os.environ.get("PORT", 8080))
    uvicorn.run(app, host="0.0.0.0", port=port)
