requirements.txt
google-adk>=0.4.0
uvicorn>=0.29.0
fastapi>=0.111.0
